requirements.txt
argon2-cffi==16.3.0
bottlenose==1.1.3
cachetools==2.0.1
Django==1.11.5
django-bootstrap-form==3.3
django-oauth-toolkit==1.0.0
django-user-accounts==2.0.3
djangorestframework==3.6.4
dj-database-url==0.4.2
dripper==1.1
pinax-theme-bootstrap==8.0.1
social-auth-app-django==1.2.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
pytz==2017.2
raven==6.2.1
requests==2.18.4
xmltodict==0.11.0
#uWSGI==2.0.15
whitenoise==3.3.1
